dist (2)/requirements.txt
playwright==1.41.1
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.23.2
websocket-client==1.6.1
requests==2.31.0 